Extract SliderBar
